Restaurar flujo original del modal y ampliar formatos permitidos
