+++ resources/views/livewire/recaudo/comunicados/create-run-modal.blade.php
@@ -70,11 +70,109 @@
                             {{ __('Cada archivo puede pesar máximo :size.', ['size' => $this->maxFileSizeLabel]) }}
                         </p>
 
-                        @livewire('recaudo.comunicados.collection-run-upload-manager', [
-                            'dataSources' => $dataSources,
-                            'maxFileSize' => $this->maxFileSizeBytes,
-                            'initialFiles' => $uploadedFiles,
-                        ], key('collection-run-upload-manager-' . ($typeId ?? 'none')))
+                        <div class="rounded-3xl border border-gray-200 bg-white shadow-sm dark:border-gray-700 dark:bg-gray-900/40">
+                            <div class="grid grid-cols-12 gap-3 border-b border-gray-200 bg-gray-100 px-4 py-3 text-xs font-semibold uppercase tracking-wide text-gray-600 dark:border-gray-700 dark:bg-gray-900 dark:text-gray-300">
+                                <div class="col-span-12 tablet:col-span-6 desktop:col-span-6">{{ __('Nombre – Código') }}</div>
+                                <div class="col-span-12 tablet:col-span-3 desktop:col-span-3">{{ __('Tipo de Archivo') }}</div>
+                                <div class="col-span-12 tablet:col-span-3 desktop:col-span-3">{{ __('Upload') }}</div>
+                            </div>
+
+                            <div class="max-h-80 overflow-y-auto divide-y divide-gray-200 px-4 dark:divide-gray-700">
+                                @foreach ($dataSources as $dataSource)
+                                    @php($fileKey = (string) ($dataSource['id'] ?? ''))
+                                    @php($selectedFile = $fileKey !== '' ? ($files[$fileKey] ?? null) : null)
+                                    <div
+                                        wire:key="data-source-{{ $dataSource['id'] }}"
+                                        class="grid grid-cols-12 items-start gap-3 py-4 text-sm text-gray-700 dark:text-gray-200"
+                                    >
+                                        <!-- Nombre y código -->
+                                        <div class="col-span-12 space-y-1 text-sm tablet:col-span-6 desktop:col-span-6">
+                                            <p class="font-semibold text-gray-900 dark:text-gray-100">
+                                                {{ $dataSource['name'] }} - {{ $dataSource['code'] }}
+                                            </p>
+                                        </div>
+
+                                        <!-- Tipo archivo -->
+                                        <div class="col-span-12 text-xs uppercase tracking-wide text-gray-600 dark:text-gray-400 tablet:col-span-3 desktop:col-span-3">
+                                            {{ $dataSource['extension'] ? strtoupper($dataSource['extension']) : __('N/A') }}
+                                        </div>
+
+                                        <!-- Uploader -->
+                                        <div class="col-span-12 space-y-2 tablet:col-span-3 desktop:col-span-3">
+                                            <div
+                                                class="space-y-2"
+                                                x-data="collectionRunUploader({
+                                                    dataSourceId: {{ $dataSource['id'] }},
+                                                    uploadUrl: '{{ route('recaudo.comunicados.uploads.chunk') }}',
+                                                    chunkSize: @js((int) config('chunked-uploads.collection_notices.chunk_size')),
+                                                    initialFile: @js(is_array($selectedFile) ? $selectedFile : null),
+                                                    maxFileSize: @js($this->maxFileSizeBytes),
+                                                })"
+                                                x-init="(() => { init(); $el.addEventListener('alpine:destroy', () => destroy(), { once: true }); })()"
+                                                :class="{ 'opacity-60': isUploading }"
+                                                wire:ignore
+                                            >
+                                                <!-- Botón seleccionar -->
+                                                <label
+                                                    for="file-{{ $dataSource['id'] }}"
+                                                    class="inline-flex w-full items-center justify-center gap-2 rounded-3xl border border-primary-300 bg-white px-4 py-2 text-sm font-semibold text-primary-900 shadow-sm transition hover:border-primary-500 hover:bg-primary-200/60 focus-within:ring-2 focus-within:ring-primary-500 focus-within:ring-offset-2 focus-within:ring-offset-white dark:border-gray-600 dark:bg-gray-900 dark:text-primary-200 dark:focus-within:ring-offset-gray-900 tablet:w-auto"
+                                                    :class="{ 'cursor-not-allowed': isUploading }"
+                                                >
+                                                    <i class="fa-solid fa-upload"></i>
+                                                    <span>{{ __('Seleccionar archivo') }}</span>
+                                                    <input
+                                                        id="file-{{ $dataSource['id'] }}"
+                                                        name="files[{{ $dataSource['id'] }}]"
+                                                        type="file"
+                                                        x-ref="fileInput"
+                                                        @change="handleFileSelected($event)"
+                                                        :disabled="isUploading"
+                                                        class="sr-only"
+                                                    />
+                                                </label>
+
+                                                <button
+                                                    type="button"
+                                                    class="inline-flex w-full items-center justify-center gap-2 rounded-3xl border border-danger bg-white px-4 py-2 text-xs font-semibold text-danger shadow-sm transition hover:bg-danger/10 focus:outline-none focus:ring-2 focus:ring-danger focus:ring-offset-2 focus:ring-offset-white disabled:cursor-not-allowed disabled:opacity-60 dark:border-danger/60 dark:bg-gray-900 dark:text-danger-200 dark:hover:bg-danger/20 dark:focus:ring-offset-gray-900 tablet:w-auto"
+                                                    x-show="isUploading || status === 'completed'"
+                                                    x-cloak
+                                                    @click="cancelUpload()"
+                                                    :disabled="isUploading && !currentSession"
+                                                >
+                                                    <i :class="isUploading ? 'fa-solid fa-ban' : 'fa-solid fa-xmark'"></i>
+                                                    <span x-text="isUploading ? '{{ __('Cancelar carga') }}' : '{{ __('Quitar archivo') }}'"></span>
+                                                </button>
+
+                                                <!-- Barra de progreso -->
+                                                <div x-show="isUploading" x-cloak class="h-2 w-full overflow-hidden rounded-full bg-gray-200 dark:bg-gray-700">
+                                                    <div
+                                                        class="h-full bg-primary-500 transition-all duration-200 ease-linear"
+                                                        :style="`width: ${progress}%`"
+                                                    ></div>
+                                                </div>
+
+                                                <!-- Nombre solo cuando termine -->
+                                                <div x-show="status === 'completed'" x-cloak>
+                                                    <div class="flex w-full items-center justify-between gap-2 rounded-3xl bg-gray-100 px-3 py-2 text-xs text-gray-700 dark:bg-gray-800 dark:text-gray-300">
+                                                        <span class="max-w-[70%] truncate" x-text="fileName"></span>
+                                                        <span class="whitespace-nowrap" x-text="progressLabel()"></span>
+                                                    </div>
+                                                </div>
+
+                                                <!-- Error -->
+                                                <p x-show="status === 'error' && errorMessage" x-cloak class="text-xs font-semibold text-danger" x-text="errorMessage"></p>
+                                            </div>
+
+                                            @error('files.' . $dataSource['id'])
+                                                <p class="inline-flex items-center rounded-3xl bg-danger px-3 py-1 text-xs font-semibold text-white">
+                                                    {{ $message }}
+                                                </p>
+                                            @enderror
+                                        </div>
+                                    </div>
+                                @endforeach
+                            </div>
+                        </div>
                     @endif
                 </div>
             </form>
@@ -85,12 +183,8 @@
 
             <div
                 class="flex w-full justify-end gap-3"
-                x-data="{
-                    canSubmit: @js($isFormValid),
-                    formReady: $wire.entangle('formReady').live,
-                }"
-                x-effect="canSubmit = Boolean(formReady)"
-                x-on:collection-run-form-state-changed.window="canSubmit = Boolean($event.detail?.isValid); formReady = canSubmit"
+                x-data="{ canSubmit: @js($isFormValid) }"
+                x-on:collection-run-form-state-changed.window="canSubmit = Boolean($event.detail?.isValid)"
             >
                 <button
                     type="button"
